Refactoriza la página de edición de perfil al mover componentes reutilizables (SectionTitle, SettingsCard, FormTile) a sus propios archivos. Esto mejora la organización del código y facilita el mantenimiento de la interfaz de usuario.
