+++ lib/presentation/widgets/atoms/section_title.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class SectionTitle extends StatelessWidget {
+  final String title;
+  const SectionTitle({super.key, required this.title});
+
+  @override
+  Widget build(BuildContext context) {
+    final ThemeData theme = Theme.of(context);
+    return Align(
+      alignment: Alignment.centerLeft,
+      child: Padding(
+        padding: const EdgeInsets.only(bottom: 8),
+        child: Text(
+          title,
+          style: theme.textTheme.labelMedium?.copyWith(color: theme.hintColor),
+        ),
+      ),
+    );
+  }
+}
