+++ lib/presentation/widgets/molecules/form_tile.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+class FormTile extends StatelessWidget {
+  final String label;
+  final Widget child;
+  const FormTile({super.key, required this.label, required this.child});
+
+  @override
+  Widget build(BuildContext context) {
+    final ThemeData theme = Theme.of(context);
+    final ColorScheme colorScheme = theme.colorScheme;
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 14),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          Expanded(
+            child: Text(
+              label,
+              style: theme.textTheme.bodyMedium?.copyWith(
+                color: colorScheme.onSurface,
+                fontWeight: FontWeight.w600,
+              ),
+            ),
+          ),
+          Flexible(child: child),
+        ],
+      ),
+    );
+  }
+}
