+++ lib/presentation/widgets/molecules/settings_card.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class SettingsCard extends StatelessWidget {
+  final List<Widget> children;
+  const SettingsCard({super.key, required this.children});
+
+  @override
+  Widget build(BuildContext context) {
+    final ThemeData theme = Theme.of(context);
+    return Container(
+      width: double.infinity,
+      decoration: BoxDecoration(
+        color: theme.colorScheme.surface,
+        borderRadius: BorderRadius.circular(16),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withValues(alpha: 0.06),
+            blurRadius: 12,
+            offset: const Offset(0, 4),
+          ),
+        ],
+      ),
+      child: Column(children: children),
+    );
+  }
+}
