+++ lib/presentation/widgets/molecules/settings_tile.dart
@@ -1,0 +1,79 @@
+import 'package:flutter/material.dart';
+
+class SettingsTile extends StatelessWidget {
+  final IconData leadingIcon;
+  final String title;
+  final Widget? trailing;
+  final VoidCallback? onTap;
+  final Color? titleColor;
+
+  const SettingsTile({
+    super.key,
+    required this.leadingIcon,
+    required this.title,
+    this.trailing,
+    this.onTap,
+    this.titleColor,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final ThemeData theme = Theme.of(context);
+    final ColorScheme colorScheme = theme.colorScheme;
+    return InkWell(
+      onTap: onTap,
+      borderRadius: BorderRadius.circular(14),
+      child: Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 14),
+        child: Row(
+          children: [
+            Container(
+              width: 36,
+              height: 36,
+              decoration: BoxDecoration(
+                color: colorScheme.secondary.withValues(alpha: 0.15),
+                borderRadius: BorderRadius.circular(10),
+              ),
+              child: Icon(leadingIcon, color: colorScheme.secondary),
+            ),
+            const SizedBox(width: 12),
+            Expanded(
+              child: Text(
+                title,
+                style: theme.textTheme.bodyMedium?.copyWith(
+                  color: titleColor ?? colorScheme.onSurface,
+                  fontWeight: FontWeight.w600,
+                ),
+              ),
+            ),
+            trailing ?? const SizedBox.shrink(),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class SettingsSwitchTile extends StatelessWidget {
+  final IconData leadingIcon;
+  final String title;
+  final bool value;
+  final ValueChanged<bool> onChanged;
+
+  const SettingsSwitchTile({
+    super.key,
+    required this.leadingIcon,
+    required this.title,
+    required this.value,
+    required this.onChanged,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return SettingsTile(
+      leadingIcon: leadingIcon,
+      title: title,
+      trailing: Switch.adaptive(value: value, onChanged: onChanged),
+    );
+  }
+}
