+++ lib/presentation/widgets/organisms/home/messages_view.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class MessagesView extends StatelessWidget {
+  const MessagesView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(child: Text('Messages view'));
+  }
+}
