+++ lib/presentation/widgets/organisms/home/store_view.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+
+class StoreView extends StatelessWidget {
+  const StoreView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final ThemeData theme = Theme.of(context);
+    final ColorScheme colorScheme = theme.colorScheme;
+    return CustomScrollView(
+      slivers: [
+        SliverToBoxAdapter(
+          child: Padding(
+            padding: const EdgeInsets.all(16),
+            child: Container(
+              height: 600,
+              decoration: BoxDecoration(
+                color: colorScheme.surface,
+                borderRadius: BorderRadius.circular(16),
+                boxShadow: [
+                  BoxShadow(
+                    color: Colors.black.withValues(alpha: 0.06),
+                    blurRadius: 12,
+                    offset: const Offset(0, 4),
+                  ),
+                ],
+              ),
+              child: const Center(child: Text('Store view')),
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
